+++ UltraPortal/Extensions/ListExtensions.cs
@@ -1,9 +1,15 @@
 using System.Collections.Generic;
+using static UltraPortal.DebugUtils;
 
 namespace UltraPortal.Extensions {
 	public static class ListExtensiosn {
 		public static void SafeAdd<T>(this List<T> list, T element) {
 			if (element == null) {
+				return;
+			}
+
+			if (list == null) {
+				LogError("List is null!");
 				return;
 			}
 
@@ -16,6 +22,11 @@ namespace UltraPortal.Extensions {
 		
 		public static void SafeRemove<T>(this List<T> list, T element) {
 			if (element == null) {
+				return;
+			}
+			
+			if (list == null) {
+				LogError("List is null!");
 				return;
 			}