+++ UltraPortal/Weapons/PortalGun.cs
@@ -99,10 +99,6 @@ namespace UltraPortal {
 		}
 
 		private void OnDestroy() {
-			if (_portal) {
-				Destroy(_portal.gameObject);
-			}
-
 			if (PortalEntry) {
 				Destroy(PortalEntry.gameObject);
 			}
@@ -110,41 +106,13 @@ namespace UltraPortal {
 			if (PortalExit) {
 				Destroy(PortalExit.gameObject);
 			}
+			
+			if (_portal) {
+				Destroy(_portal.gameObject);
+			}
 		}
 
 		private void InitPortals() {
-			// _portalObject = new GameObject("Portal") {
-			// 	layer = PortalLayer
-			// };
-   //
-			// _portal = _portalObject.AddComponent<Portal>(); 
-			//
-			// _portal.additionalSampleThreshold = 0;
-			// _portal.allowCameraTraversals = true;
-			// _portal.appearsInRecursions = true;
-			// _portal.canHearAudio = false;
-			// _portal.canSeeItself = true;
-			// _portal.canSeePortalLayer = true;
-			// _portal.clippingMethod = PortalClippingMethod.Default;
-			// _portal.consumeAudio = false;
-			// _portal.disableRange = 0;
-			// _portal.enableOverrideFog = false;
-			// _portal.enterOffset = 1.5f;
-			// _portal.entry = PortalEntry.transform;
-   //          _portal.minimumEntrySideSpeed = ModConfig.MinimumEntryExitSpeed;
-   //          
-			// _portal.exit = PortalExit.transform;
-			// _portal.exitOffset = 1.5f;
-			// _portal.minimumExitSideSpeed = ModConfig.MinimumEntryExitSpeed;
-			//
-			// _portal.renderSettings = PortalSideFlags.Enter | PortalSideFlags.Exit;
-			// _portal.fakeVPMatrix = Matrix4x4.zero;
-			// _portal.mirror = false;
-			// _portal.shape = new PlaneShape {
-			// 	width = _portalSize.x,
-			// 	height = _portalSize.y
-			// };
-
 			_portal = CreatePortal("Portal", PortalEntry.transform, PortalExit.transform, _portalSize);
 		}